refactored admin login to not require token
