+++ app/routers/LoginRouter.js
@@ -1,6 +1,14 @@
+var authenticate = require ('../middleware/authenticate');
+
 // UserRouter
 module.exports = exports = {
   '/login': {
     post: { action: 'LoginController@login' }
+  },
+  '/admin': {
+    '/login': {
+      use: authenticate.isAdminUser,
+      post: { action: 'LoginController@login' }
+    }
   }
 };
