--- app/routers/v1/admin/LoginRouter.js
@@ -1,6 +1,0 @@
-// UserRouter
-module.exports = exports = {
-  '/login': {
-    post: { action: 'LoginController@login' }
-  }
-};
